feat(import): 添加整合包导入核心逻辑及相关功能
