+++ src-tauri/src/adapter/tauri/commands/instance.rs
@@ -7,17 +7,16 @@
 //! 错误统一为接口契约错误 [`InstanceServiceError`]，可序列化回前端，
 //! 不携带底层敏感细节。
 
-use std::path::PathBuf;
 use std::sync::Arc;
-use std::time::{SystemTime, UNIX_EPOCH};
 
 use sealantern_application::error::InstanceError;
 use sealantern_application::service::CoreInstanceService;
 use sealantern_application::services::AppServices;
-use sealantern_core::instance::{Instance, InstanceId, InstanceSpec, LocalLaunch, StartupMode};
+use sealantern_core::instance::{Instance, InstanceId, InstanceSpec};
 use sealantern_core::provisioning::{
     ImportExistingServerError as CoreImportError, ImportExistingServerRequest,
-    SourceDirectoryError, build_import_spec, plan_existing_instance, source_directories_equal,
+    ImportModpackRequest, SourceDirectoryError, SourceType, build_import_spec, infer_source_type,
+    plan_existing_instance, plan_import_modpack, source_directories_equal,
     validate_source_directory,
 };
 use sealantern_infra::archive::extract_zip;
@@ -175,72 +174,6 @@ fn import_error(code: &'static str, message: impl Into<String>) -> ImportExistin
     }
 }
 
-// ============================================================================
-// import_modpack 命令实现
-// ============================================================================
-
-/// 来源类型枚举。
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum SourceType {
-    /// zip/tar.gz/tgz 压缩包。
-    Archive,
-    /// 单个 jar 文件。
-    JarFile,
-    /// 已存在的文件夹。
-    Folder,
-}
-
-/// 根据路径扩展名推断来源类型。
-fn infer_source_type(path: &std::path::Path) -> SourceType {
-    let path_str = path.to_string_lossy().to_lowercase();
-
-    // .tar.gz 和 .tgz 必须先判断
-    if path_str.ends_with(".tar.gz") || path_str.ends_with(".tgz") {
-        return SourceType::Archive;
-    }
-
-    let ext = path
-        .extension()
-        .and_then(|s| s.to_str())
-        .map(|s| s.to_lowercase());
-
-    match ext.as_deref() {
-        Some("zip") => SourceType::Archive,
-        Some("jar") => SourceType::JarFile,
-        _ => SourceType::Folder,
-    }
-}
-
-/// 整合包导入请求。
-#[derive(Debug, Clone, serde::Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub struct ImportModpackRequest {
-    /// 实例名称。
-    pub name: String,
-    /// 整合包来源（zip 路径或文件夹路径）。
-    pub modpack_path: PathBuf,
-    /// Java 可执行文件路径。
-    pub java_path: PathBuf,
-    /// 最大内存（MiB）。
-    pub max_memory: u32,
-    /// 最小内存（MiB）。
-    pub min_memory: u32,
-    /// 监听端口。
-    pub port: u16,
-    /// 启动模式。
-    pub startup_mode: String,
-    /// 启动目标文件路径（相对于 run_path）。
-    pub startup_file_path: Option<PathBuf>,
-    /// 核心类型（paper、forge 等）。
-    pub core_type: Option<String>,
-    /// Minecraft 版本。
-    pub mc_version: Option<String>,
-    /// 自定义启动命令。
-    pub custom_command: Option<String>,
-    /// 目标运行目录。
-    pub run_path: PathBuf,
-}
-
 /// 整合包导入失败时返回给前端的错误。
 #[derive(Debug, serde::Serialize)]
 pub struct ImportModpackError {
@@ -266,61 +199,6 @@ fn modpack_error(code: &'static str, message: impl Into<String>) -> ImportModpac
     }
 }
 
-/// 生成实例 ID（时间戳毫秒）。
-fn generate_instance_id() -> String {
-    let ts = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|d| d.as_millis())
-        .unwrap_or(0);
-    format!("{}", ts)
-}
-
-/// 构建 InstanceSpec。
-fn build_instance_spec(
-    request: &ImportModpackRequest,
-    directory: &std::path::Path,
-    startup_target: Option<PathBuf>,
-) -> Result<InstanceSpec, ImportModpackError> {
-    let id = generate_instance_id();
-    let now = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|d| d.as_secs())
-        .unwrap_or(0);
-
-    let startup_mode = StartupMode::parse(&request.startup_mode).map_err(|error| {
-        modpack_error("invalid_startup_mode", format!("invalid startup mode: {error}"))
-    })?;
-
-    let id = InstanceId::new(id).map_err(|error| {
-        modpack_error("invalid_instance_id", format!("failed to create instance ID: {error}"))
-    })?;
-
-    Ok(InstanceSpec {
-        id,
-        name: request.name.clone(),
-        aliases: Vec::new(),
-        core_type: request.core_type.clone().unwrap_or_default(),
-        core_version: String::new(),
-        game_version: request.mc_version.clone().unwrap_or_default(),
-        directory: directory.to_path_buf(),
-        port: request.port,
-        max_memory_mib: request.max_memory,
-        min_memory_mib: request.min_memory,
-        created_at_unix_secs: now,
-        last_started_at_unix_secs: None,
-        server_metadata: None,
-        launch: LocalLaunch {
-            startup_mode,
-            startup_target,
-            custom_command: request.custom_command.clone(),
-            custom_executable: None,
-            custom_arguments: Vec::new(),
-            java_executable: Some(request.java_path.clone()),
-            jvm_arguments: Vec::new(),
-        },
-    })
-}
-
 /// 导入整合包或服务器文件夹。
 ///
 /// 支持三种来源：
@@ -329,60 +207,57 @@ fn build_instance_spec(
 /// - 文件夹：直接引用原路径
 #[tauri::command(rename_all = "snake_case")]
 pub async fn import_modpack(request: ImportModpackRequest) -> Result<Instance, ImportModpackError> {
+    use sealantern_core::provisioning::ImportModpackError as CoreError;
+
     // 1. 判断来源类型
     let source_type = infer_source_type(&request.modpack_path);
 
-    // 2. 根据类型处理文件
-    let (effective_directory, startup_target) = match source_type {
+    // 2. 规划导入（构建 InstanceSpec）
+    let result = plan_import_modpack(&request).map_err(|error| match error {
+        CoreError::InvalidStartupMode(msg) => {
+            modpack_error("invalid_startup_mode", format!("invalid startup mode: {msg}"))
+        }
+        CoreError::InvalidInstanceId(err) => {
+            modpack_error("invalid_instance_id", format!("failed to create instance ID: {err}"))
+        }
+        CoreError::ExtractFailed(msg) => modpack_error("extract_failed", msg),
+        CoreError::CreateDirectoryFailed(msg) => modpack_error("create_directory_failed", msg),
+        CoreError::CopyFailed(msg) => modpack_error("copy_failed", msg),
+    })?;
+
+    // 3. 执行文件操作（根据来源类型）
+    match source_type {
         SourceType::Archive => {
             // 解压 zip/tar.gz 到 run_path
-            extract_zip(&request.modpack_path, &request.run_path).map_err(|error| {
+            extract_zip(&request.modpack_path, &result.directory).map_err(|error| {
                 modpack_error("extract_failed", format!("failed to extract archive: {error}"))
             })?;
-            let target = request
-                .startup_file_path
-                .as_ref()
-                .map(|p| request.run_path.join(p));
-            (request.run_path.clone(), target)
         }
         SourceType::JarFile => {
             // 创建 run_path 目录并复制 jar
-            std::fs::create_dir_all(&request.run_path).map_err(|error| {
+            std::fs::create_dir_all(&result.directory).map_err(|error| {
                 modpack_error(
                     "create_directory_failed",
                     format!("failed to create run directory: {error}"),
                 )
             })?;
-            let jar_name = request
-                .modpack_path
-                .file_name()
-                .and_then(|n| n.to_str())
-                .unwrap_or("server.jar");
-            let dest_path = request.run_path.join(jar_name);
-            std::fs::copy(&request.modpack_path, &dest_path).map_err(|error| {
-                modpack_error("copy_jar_failed", format!("failed to copy jar file: {error}"))
-            })?;
-            (request.run_path.clone(), Some(dest_path))
+            if let Some(ref dest_path) = result.startup_target {
+                std::fs::copy(&request.modpack_path, dest_path).map_err(|error| {
+                    modpack_error("copy_jar_failed", format!("failed to copy jar file: {error}"))
+                })?;
+            }
         }
         SourceType::Folder => {
-            // 直接引用原目录
-            let target = request
-                .startup_file_path
-                .as_ref()
-                .map(|p| request.modpack_path.join(p));
-            (request.modpack_path.clone(), target)
+            // 直接引用原目录，无需文件操作
         }
-    };
-
-    // 3. 构建 InstanceSpec
-    let spec = build_instance_spec(&request, &effective_directory, startup_target)?;
+    }
 
     // 4. 创建实例
     let service = instance_service()
         .await
         .map_err(|error| modpack_error("service_unavailable", error.to_string()))?;
     service
-        .create(spec)
+        .create(result.spec)
         .await
         .map_err(|error| modpack_error("create_failed", error.to_string()))
 }
